\name{Cloglin_mult}
\alias{Cloglin_mult}
\title{The causal log-linear model}
\description{ It calculates the causal parameters for a causal log-linear model with multiplicative interaction
}
\value{ The estimated causal parameters of the additive form, their standard errors
and their p-values which test H_0: parameter=0.
}
\references{
Gheno, G. (2015), The causal analysis in the loglinear model}
\author{
Gloria Gheno

Maintainer: Gloria Gheno <gloriagheno@libero.it>
}

