\name{efflog-package}
\alias{efflog-package}
\alias{efflog}
\docType{package}
\title{The Causal Effects for a Causal Loglinear Model
}
\description{ Calculate the Causal Effects for a causal loglinear model with the multiplicative interaction or without the multiplicative interaction}
\details{
\tabular{ll}{
Package: \tab efflog\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-14\cr
License: \tab GPL (>=2)\cr
}}
\author{
Gloria Gheno <gloriagheno@libero.it>
}
\references{Gheno Gloria (2015), The causal analysis in the loglinear model}
\keyword{ package }

