\name{exp_par}
\alias{exp_par}
\title{The causal log-linear model}
\description{ It calculates the causal parameters for a causal log-linear model
}
\value{ The estimated causal parameters of the multiplicative form
}
\references{
Gheno, G. (2015), The causal analysis in the loglinear model}
\author{
Gloria Gheno

Maintainer: Gloria Gheno <gloriagheno@libero.it>
}

