\name{exp_par_mult}
\alias{exp_par_mult}
\title{The causal log-linear model}
\description{ It calculates the causal parameters for a causal log-linear model with the multiplicative interaction
}
\value{ The estimated causal parameters of the multiplicative form
}
\references{
Gheno, G. (2015), The causal analysis in the loglinear model}
\author{
Gloria Gheno

Maintainer: Gloria Gheno <gloriagheno@libero.it>
}

